Move components into their own file
